[backend] max shareable markings check for public dashboard (#7006)
