+++ opencti-platform/opencti-graphql/src/modules/publicDashboard/publicDashboard-domain.ts
@@ -41,10 +41,11 @@ import {
 } from '../../domain/stixCoreObject';
 import { ABSTRACT_STIX_CORE_OBJECT } from '../../schema/general';
 import { findAll as stixRelationships, stixRelationshipsDistribution, stixRelationshipsMultiTimeSeries, stixRelationshipsNumber } from '../../domain/stixRelationship';
-import { bookmarks, getAvailableDataSharingMarkings } from '../../domain/user';
+import { bookmarks, checkUserCanShareMarkings, getAvailableDataSharingMarkings } from '../../domain/user';
 import { daysAgo } from '../../utils/format';
 import { isStixCoreObject } from '../../schema/stixCoreObject';
 import { ES_MAX_CONCURRENCY } from '../../database/engine';
+import { findById as findMarkingDefinitionById } from '../../domain/markingDefinition';
 
 export const findById = (
   context: AuthContext,
@@ -127,7 +128,14 @@ export const addPublicDashboard = async (
 
   const access = getUserAccessRight(user, dashboard);
   if (access !== MEMBER_ACCESS_RIGHT_ADMIN) {
-    throw ForbiddenAccess('You are not allowed to do this.');
+    throw ForbiddenAccess();
+  }
+
+  if (input.allowed_markings_ids && input.allowed_markings_ids.length > 0) {
+    const markingLevels = await Promise.all(input.allowed_markings_ids.map((id) => {
+      return findMarkingDefinitionById(context, user, id);
+    }));
+    await checkUserCanShareMarkings(context, user, markingLevels);
   }
 
   const uriKey = input.uri_key.replace(/[^a-zA-Z0-9\s-]+/g, '').replace(/\s+/g, '-');
